refactor(core): extract response.rs from main.rs
Pure code move: 22 response constructors and small header helpers
relocate from main.rs into a new core/src/response.rs. No logic
change. Function bodies are byte-identical to what they replaced.

Moved:
  basic errors:
    not_found, unauthorized, bad_request, payload_too_large,
    precondition_failed, server_error, method_not_allowed,
    options_response
  storage / quota:
    insufficient_storage, storage_quota_exceeded, storage_error,
    is_insufficient_storage_error
  audit verify:
    audit_valid, audit_broken, audit_not_applicable,
    audit_header_value
  proc helpers:
    proc_text_response, du_body, df_body, world_list_body
  header utility:
    to_header_map

All marked pub(crate). main.rs adds `mod response;` and
`pub(crate) use crate::response::*;` so existing call sites in
http_semantics.rs / listen.rs / coap.rs (which reference these as
`crate::not_found` etc.) keep working with no import churn.

Stays in main.rs:
- BlockingSqliteError enum and blocking_storage_error mapper
  (tied to the spawn_blocking scaffold, not the response surface)
- exact_or_child, can_read, apply_meta_headers
  (Core-shaped helpers, not pure response constructors)

This is the first PR of the v7 FSM-pipeline cascade. Base branch is
fix/per-world-lock, not master, so the diff shows only the
extraction. PR 2 (path.rs) will branch off this.

Stats:
  336 insertions (response.rs + main.rs structural)
  291 deletions (old function bodies removed from main.rs)
  Pure mv: every inserted body is byte-identical to a deletion.

cargo fmt --manifest-path core/Cargo.toml -- --check: clean
cargo clippy --manifest-path core/Cargo.toml --all-targets -- -D warnings: clean
cargo test --manifest-path core/Cargo.toml: 100 passed, 0 failed
